Move driri.py to addons_compat/ (pure addon translator)
DateRangeInRangeIndex is a third-party addon (collective.DateRangeInRangeIndex),
unlike DateRecurringIndex which is a Plone core dependency. Move its translator
to the addons_compat/ package where it belongs alongside eea.facetednavigation.
